fix stackoverflow error after lambda parents fix & fix another progression test
